Vote time processor

+++ lib/processors/device_model_processor.rb
@@ -1,7 +1,6 @@
-require "seven_zip_ruby"
-require_relative "common/constants.rb"
+require_relative "../common/constants.rb"
 
-class DeviceModelCollector
+class DeviceModelProcessor
 
   MODEL_FIELD = "Modelo de Urna: "
 
@@ -14,18 +13,16 @@ class DeviceModelCollector
   end
 
   def get_device_model_from_location(estado, cod_municipio, zona, secao, cod_turno)
-    logjez_file_path = LogsDownloader.new(workspace, estado, cod_municipio, zona, secao, cod_turno).fetch_logs
+    logs_downloader = LogsDownloader.new(workspace, estado, cod_municipio, zona, secao, cod_turno)
+    logjez_file_path = logs_downloader.fetch_logs
+    log_path = LogsDownloader.extract_logjez_file(logjez_file_path, tmp)
 
-    result = get_device_model_from_zip(logjez_file_path)
-
-    result
+    get_device_model(log_path)
   end
 
   # Returns the device model given the *.logjez compressed file
   def get_device_model_from_zip(logjez_file)
-    extract_logjez_file(logjez_file, tmp)
-    
-    dest_path = File.join(tmp, "logd.dat")
+    dest_path = LogsDownloader.extract_logjez_file(logjez_file, tmp)
     result = get_device_model(dest_path)
 
     File.delete(dest_path) if File.exist?(dest_path)
@@ -35,6 +32,8 @@ class DeviceModelCollector
 
   # Returns the device model given the logd.dat log text file
   def get_device_model(log_file)
+    return "-" unless File.exist?(log_file)
+
     File.open(log_file, encoding: ENCODING) do |file|
       line = file.find { |line| line.include?(MODEL_FIELD) }
       start_ix = line.index(MODEL_FIELD) + MODEL_FIELD.size
@@ -42,11 +41,4 @@ class DeviceModelCollector
     end
   end
 
-  private 
-
-  def extract_logjez_file(logjez_file, dest_path)
-    File.open(logjez_file, "rb") do |file|
-      SevenZipRuby::Reader.extract_all(file, dest_path)
-    end
-  end
 end
